Major Changes:
Modified the item controller to contain an item of reference, as well as
the user controller to have a user of reference. Removed usernameString
from majority of code as it wasn't needed for intents with this new
process.

Added back "cancel" button for edit item, as owner may wish to cancel an
edit.